watcher/requirements.txt
watchdog>=4.0.0
requests>=2.31.0
lz4>=4.3.0
pycryptodome>=3.20.0
